Rename .js files to .jsx
