+++ .eslintrc
@@ -11,9 +11,6 @@
     // import/first is incompatible with airbnb config
     "import/first": 0,
 
-    // allow files to have extension ".js"
-    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-
     "react/prop-types": [2, { "skipUndeclared": true }],
 
     // allow ommitted parentheses in single-argument arrow functions
